# server/src/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/request-logs/statistics": {
      "get": {
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/admin/request-logs/statistics"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get request log statistics",
        "tags": ["Request Logs"]
      }
    },
...
    "/api/v1/users/statistics": {
      "get": {
        "description": "Retrieves user statistics including total, active, inactive, and today registered counts. Requires Admin or Moderator role.",
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User statistics retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "statusCode": 200,
                    "message": "Statistics fetched successfully",
                    "data": {
                      "total": 125,
                      "active": 98,
                      "inactive": 27,
                      "pending": 0,
                      "todayRegistered": 5
                    },
                    "meta": {
                      "cached": false,
                      "timestamp": "2026-01-31T10:00:00.000Z"
                    },
                    "timestamp": "2026-01-31T10:00:00.015Z",
                    "path": "/api/v1/users/statistics"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/users/statistics"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/users/statistics"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user statistics (Admin/Moderator)",
        "tags": ["users"]
      }
    },
...
    "/api/v1/roles/statistics": {
      "get": {
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role statistics fetched successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/roles/statistics"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get role statistics",
        "tags": ["RBAC - Roles"]
      }
    },
...
    "/api/v1/permissions/statistics": {
      "get": {
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission statistics fetched successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/permissions/statistics"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get permission statistics",
        "tags": ["RBAC - Permissions"]
      }
    },
...
    "/api/v1/newsletter/statistics": {
      "get": {
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newsletter statistics fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "statusCode": 200,
                    "message": "Newsletter statistics fetched successfully",
                    "data": {
                      "total": 1000,
                      "active": 850,
                      "inactive": 150,
                      "todaySubscribed": 25
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/newsletter/statistics"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get newsletter statistics (admin)",
        "tags": ["newsletter"]
      }
    },
...
    "/api/v1/contact/statistics": {
      "get": {
        "description": "Admin endpoint to get contact submission statistics",
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "statusCode": 200,
                    "message": "Contact statistics fetched successfully",
                    "data": {
                      "total": 150,
                      "new": 25,
                      "read": 50,
                      "replied": 60,
                      "today": 5
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/contact/statistics"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get contact statistics",
        "tags": ["contact"]
      }
    },
...
    "/api/v1/blog": {
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "statusCode": 201,
                    "message": "Blog post created successfully",
                    "data": {
                      "id": "uuid",
                      "title": "How to Build a Modern Web App",
                      "slug": "how-to-build-a-modern-web-app",
                      "content": "<h1>Introduction</h1>",
                      "status": "draft",
                      "readTime": 5,
                      "createdAt": "2026-02-02T10:00:00.000Z"
                    },
                    "meta": {
                      "blog_post_id": "uuid",
                      "slug": "how-to-build-a-modern-web-app",
                      "created_at": "2026-02-02T10:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 400,
                    "message": "Validation failed",
                    "errors": ["Field is required", "Invalid format"],
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new blog post",
        "tags": ["blog"]
      }
    },
    "/api/v1/blog/admin": {
      "get": {
        "operationId": "findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog posts fetched successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog/admin"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all blog posts (admin)",
        "tags": ["blog"]
      }
    },
    "/api/v1/blog/published": {
      "get": {
        "operationId": "findPublished",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published blog posts fetched successfully"
          }
        },
        "summary": "Get published blog posts (public)",
        "tags": ["blog"]
      }
    },
    "/api/v1/blog/slug/{slug}": {
      "get": {
        "operationId": "findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog post fetched successfully"
          },
          "404": {
            "description": "Not Found - Blog post not found.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 404,
                    "message": "Blog post not found",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog/slug/:slug"
                  }
                }
              }
            }
          }
        },
        "summary": "Get blog post by slug (public)",
        "tags": ["blog"]
      }
    },
    "/api/v1/blog/statistics": {
      "get": {
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog statistics fetched successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog/statistics"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get blog statistics",
        "tags": ["blog"]
      }
    },
    "/api/v1/blog/categories": {
      "get": {
        "operationId": "getCategories",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categories fetched successfully"
          }
        },
        "summary": "Get all blog categories with post counts",
        "tags": ["blog"]
      },
      "post": {
        "operationId": "createCategory",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully"
          },
          "400": {
            "description": "Bad Request - Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 400,
                    "message": "Validation failed",
                    "errors": ["Field is required", "Invalid format"],
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog/categories"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 401,
                    "message": "Unauthorized",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog/categories"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "error",
                    "statusCode": 409,
                    "message": "Category already exists",
                    "timestamp": "2026-01-31T19:20:00.000Z",
                    "path": "/api/v1/blog/categories"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new blog category with SEO fields",
        "tags": ["blog"]
      }
    },
...
    "/api/v1/blog/tags": {
      "get": {
        "operationId": "getTags",
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version for header-based versioning",
            "schema": {
              "type": "string",
              "enum": ["1", "2"],
              "default": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tags fetched successfully"
          }
        },
        "summary": "Get all blog tags",
        "tags": ["blog"]
      }
    },
...
  },
  "info": {
    "title": "Fullstack NNP API",
...
    "version": "1.0.0",
    "contact": {
      "name": "K K Kaundal",
      "url": "https://github.com/k-kaundal",
      "email": "kk@kaundal.vip"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development"
    },
    {
      "url": "https://your-api.vercel.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter your JWT access token",
        "in": "header"
      },
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for external integrations (optional)"
      }
    },
    "schemas": {
...
      "CreateBlogCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Technology",
            "description": "Category name"
          },
          "description": {
            "type": "string",
            "example": "Latest technology news and tutorials",
            "description": "Category description"
          },
          "image": {
            "type": "string",
            "example": "https://example.com/images/tech.jpg",
            "description": "Category image URL"
          },
          "metaDescription": {
            "type": "string",
            "example": "Explore the latest in technology",
            "description": "Meta description for SEO"
          },
          "metaKeywords": {
            "type": "string",
            "example": "technology, tech news, tutorials",
            "description": "Meta keywords for SEO"
          }
        },
        "required": ["name"]
      },
...
    }
  }
}